Add WDL file, start batch process script
